chore: rename stress to *ignored_test.rs
